+++ lib/src/observer/controller_observer.dart
@@ -1,49 +1,50 @@
-import 'dart:async';
+import 'dart:developer';
 
 import 'package:meta/meta.dart';
 import 'package:streamed_controller/src/controller.dart';
 
-// typedef ControllerEvent =
-
 /// ControllerObserver Singleton class
-class BaseControllerObserver {
-  static final BaseControllerObserver _internalSingleton =
-      BaseControllerObserver._internal();
-  factory BaseControllerObserver() => _internalSingleton;
+class StreamedControllerObserver {
+  @protected
+  StreamedControllerObserver();
 
-  BaseControllerObserver._internal() {
-    _$zone = Zone.current.fork(
-        specification: ZoneSpecification(
-            handleUncaughtError: (_, __, ___, error, stackTrace) =>
-                onError(null, error, stackTrace)));
-  }
-
-  late final Zone _$zone;
+  factory StreamedControllerObserver.dartLog() => _TestControllerObserver();
 
   static Never _overrideError() => throw UnimplementedError(
       'This method should be overrided with your own logging package/solution');
 
-  FutureOr<void> handleError(FutureOr<void> action) async => _$zone.runGuarded(
-        () async => action,
-      );
+  @mustBeOverridden
+  void onCreate(StreamedController controller) => _overrideError();
 
   @mustBeOverridden
-  void onCreate(BaseStreamedController controller) => _overrideError();
+  void onDispose(StreamedController controller) => _overrideError();
 
   @mustBeOverridden
-  void onDispose(BaseStreamedController controller) => _overrideError();
-
-  @mustBeOverridden
-  void onStateChanged(BaseStreamedController controller, Object prevState,
-          Object nextState) =>
+  void onStateChanged(
+          StreamedController controller, Object prevState, Object nextState) =>
       _overrideError();
 
   @mustBeOverridden
-  void onError(BaseStreamedController? controller, Object error,
+  void onError(StreamedController? controller, Object? error,
           StackTrace stackTrace) =>
       _overrideError();
 }
 
-class StreamedControllerObserver {
-  static BaseControllerObserver? observer;
+class _TestControllerObserver extends StreamedControllerObserver {
+  @override
+  void onCreate(StreamedController<Object> controller) =>
+      log('${controller.runtimeType} created');
+
+  @override
+  void onDispose(StreamedController<Object> controller) =>
+      log('${controller.runtimeType} disposed');
+  @override
+  void onError(StreamedController<Object>? controller, Object? error,
+          StackTrace stackTrace) =>
+      log('${controller.runtimeType} got error: $error');
+
+  @override
+  void onStateChanged(StreamedController<Object> controller, Object prevState,
+          Object nextState) =>
+      log('${controller.runtimeType} state updated, new state: $nextState');
 }